backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
pydantic==2.9.2
email-validator==2.2.0
psycopg2-binary==2.9.10
